backend/requirements.txt
# python 3.11
alibabacloud_ecs20140526==3.0.8
alibabacloud_tea_openapi==0.3.7
alibabacloud_tea_console==0.0.1
alibabacloud_tea_util==0.3.11
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
markdown==3.4.4
django-filter==23.2
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
django-apscheduler==0.6.2
psycopg2==2.9.7